test(zephyr-metro-plugin): relocate test files to dedicated __test__ directory
- Moved integration and unit test files for better organization.
- Updated imports to reflect new directory structure.
